requirements.txt
aldryn-apphooks-config==0.3.3
aldryn-background-image==0.1.1
aldryn-boilerplates==0.7.5
aldryn-bootstrap3==1.2.0
aldryn-categories==1.0.4
aldryn-common==1.0.4
aldryn-forms==2.1.3
-e git+https://github.com/aldryn/aldryn-newsblog.git@d3cbab21df189e55d20dc696e2733de6f40c1ae7#egg=aldryn_newsblog
aldryn-people==1.2.2
aldryn-reversion==1.1.0
aldryn-translation-tools==0.2.1
Babel==2.3.4
backport-collections==0.1
cmsplugin-filer==1.1.3
cssselect==1.0.1
cssutils==1.0.2
dj-database-url==0.4.2
Django==1.8.17
django-absolute==0.3
django-appconf==1.0.2
django-appdata==0.1.6
django-classy-tags==0.8.0
django-cms==3.4.2
django-durationfield==0.5.3
django-emailit==0.2.2
django-filer==1.2.7
django-formtools==2.0
django-mptt==0.8.7
django-parler==1.7
django-phonenumber-field==1.1.0
django-polymorphic==1.0.2
django-reversion==1.10.2
django-sekizai==0.10.0
django-simple-captcha==0.5.3
django-sizefield==0.9.1
django-sortedm2m==1.3.3
django-tablib==3.1.2
django-taggit==0.22.0
django-treebeard==4.1.0
djangocms-admin-style==1.2.6.2
djangocms-attributes-field==0.2.0
djangocms-cascade==0.12.0
djangocms-column==1.7.0
djangocms-googlemap==1.0.2
djangocms-installer==0.9.5
djangocms-link==2.1.1
djangocms-snippet==1.9.2
djangocms-style==2.0.1
djangocms-text-ckeditor==3.4.0
djangocms-video==2.0.3
easy-thumbnails==2.3
html5lib==0.9999999
jsonfield==2.0.0
lxml==3.7.3
olefile==0.44
phonenumbers==8.3.2
phonenumberslite==8.3.2
pillow==4.0.0
premailer==3.0.1
python-dateutil==2.6.0
python-slugify==1.2.0
pytz==2016.10
requests==2.13.0
six==1.10.0
tablib==0.11.4
tzlocal==1.3
Unidecode==0.4.20
YURL==0.13
